From Alan:
This adds a config option "bindaddr" that forces vtund to listen on a
particular address in server mode. The syntax is identical to the
existing srcaddr option (i.e. you can bind to a device name, an ip, or
hostname) and appears in the general section of the config file. If
bindaddr is omitted, vtund listens on 0.0.0.0, which is the current
behavior.

Factored the "fill in a sockaddr from dev/ip/hostname" stuff out of
local_addr() and into a new function generic_addr() so bindaddr can
use it too. Moved the svr_port member of the vtun struct into a new
vtun_addr member which holds both listen port and listen address
now. Syntax for "port" remains the same though...so there are this and
other inconsistencies in the config semantics now, but I didn't feel
comfortable fixing them.